test(use-popper): migrate hook tests to browser mode
